+++ src/N98/Magento/Command/AbstractMagentoCommand.php
@@ -345,7 +345,13 @@ abstract class AbstractMagentoCommand extends Command
     public function getComposer(InputInterface $input, OutputInterface $output)
     {
         $io = new ConsoleIO($input, $output, $this->getHelperSet());
-        return ComposerFactory::create($io, array());
+        $config = array(
+            'config' => array(
+                'secure-http' => false,
+            ),
+        );
+
+        return ComposerFactory::create($io, $config);
     }
 
     /**